+++ config/config.js
@@ -4,6 +4,7 @@ const rootPath = path.normalize(`${__dirname}/..`);
 
 module.exports = {
   analytics: {
+    adobeDatalayerUrl: process.env.ADOBE_DATALAYER_URL,
     adobeTrackingUrl: process.env.ADOBE_TRACKING_URL,
     googleAnalyticsId: process.env.GOOGLE_ANALYTICS_TRACKING_ID,
     hotjarId: process.env.HOTJAR_ANALYTICS_TRACKING_ID,