+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-0.19.0 / 2018-11-TBD
+0.19.0 / 2018-11-20
 ===================
 - Update npm dependencies
 - Create `digitalData` object for use by Adobe Analytics
+- Update CSP for compatibility with Adobe Analytics
 
 0.18.0 / 2018-10-09
 ===================